separate messages from gnss_comm
