+++ gnss_comm_msgs/msg/GnssTimeMsg.msg
